+++ src/Resources/contao/languages/de/default.php
@@ -1,5 +1,6 @@
 <?php
 
 $GLOBALS['TL_LANG']['MSC']['contaoMediaLibraryBundle'] = [
-    'messageOriginalFileKept' => 'Das Produkt wurde erfolgreich gelöscht. Achtung: die Originaldatei ist immer noch in der Dateiverwaltung vorhanden, um Datenverlust zu vermeiden.'
+    'messageOriginalFileKept' => 'Das Produkt wurde erfolgreich gelöscht. Achtung: die Originaldatei ist immer noch in der Dateiverwaltung vorhanden, um Datenverlust zu vermeiden.',
+    'additional' => 'zusätzlich'
 ];