+++ src/Resources/contao/languages/de/tl_ml_download.php
@@ -5,13 +5,14 @@ $lang = &$GLOBALS['TL_LANG']['tl_ml_download'];
 /**
  * Fields
  */
-$lang['tstamp']    = ['Änderungsdatum', ''];
-$lang['title']     = ['Titel', 'Geben Sie hier bitte den Titel ein.'];
-$lang['published'] = ['Veröffentlichen', 'Wählen Sie diese Option zum Veröffentlichen.'];
-$lang['start']     = ['Anzeigen ab', 'Download erst ab diesem Tag auf der Webseite anzeigen.'];
-$lang['stop']      = ['Anzeigen bis', 'Download nur bis zu diesem Tag auf der Webseite anzeigen.'];
-$lang['file']      = ['Datei', 'Laden Sie hier die Download-Datei hoch.'];
-$lang['imageSize'] = ['Bildgröße', 'Der Download liegt in o.g. Bildgröße vor.'];
+$lang['tstamp']       = ['Änderungsdatum', ''];
+$lang['title']        = ['Titel', 'Geben Sie hier bitte den Titel ein.'];
+$lang['file']         = ['Datei', 'Laden Sie hier die Download-Datei hoch.'];
+$lang['imageSize']    = ['Bildgröße', 'Der Download liegt in o.g. Bildgröße vor.'];
+$lang['isAdditional'] = ['Gehört zu einer zusätzlichen Datei', 'Der Download gehört zu einer zusätzlichen Datei.'];
+$lang['published']    = ['Veröffentlichen', 'Wählen Sie diese Option zum Veröffentlichen.'];
+$lang['start']        = ['Anzeigen ab', 'Download erst ab diesem Tag auf der Webseite anzeigen.'];
+$lang['stop']         = ['Anzeigen bis', 'Download nur bis zu diesem Tag auf der Webseite anzeigen.'];
 
 /**
  * Legends